move up script details rendering to fix UI bug
